+++ astropy/units/physical.py
@@ -30,6 +30,68 @@ class PhysicalType:
     """
     Represents and provides information on the physical type(s) that are
     associated with a set of units.
+
+    Parameters
+    ----------
+    unit : u.Unit
+        The unit to be represented by the physical type.
+
+    physical_types : `str` or `set` of `str`
+        A `str` representing the name of the physical type of the unit,
+        or a `set` containing strings that represent one or more names
+        of physical types.
+
+    Examples
+    --------
+    The preferred method to access a physical type is by accessing the
+    ``physical_type`` attribute of a unit.
+
+    >>> import astropy.units as u
+    >>> u.meter.physical_type
+    'length'
+
+    Occasionally, a set of units will correspond to more than one
+    physical type.
+
+    >>> u.Pa.physical_type
+    {'energy density', 'pressure'}
+    >>> u.Pa.physical_type == 'pressure'
+    True
+    >>> 'energy density' in u.Pa.physical_type
+    True
+
+    Physical types may be used for dimensional analysis.
+
+    >>> length = u.pc.physical_type
+    >>> area = (u.cm ** 2).physical_type
+    >>> length * area
+    'volume'
+
+    Multiplcation, division, and exponentiation work with physical types.
+
+    >>> length * area
+    'volume'
+    >>> area / length
+    'length'
+    >>> length ** 3
+    'volume'
+
+    Unknown physical types will be labelled as ``"unknown"``.
+
+    >>> (u.s ** 13).physical_type
+    'unknown'
+
+    Dimensional analysis may be performed for unknown physical types too.
+
+    >>> length_to_19th_power = (u.m ** 19).physical_type
+    >>> length_to_20th_power = (u.m ** 20).physical_type
+
+    >>> length_to_19th_power, length_to_20th_power
+    ('unknown', 'unknown')
+
+    >>> length_to_20th_power / length_to_19th_power
+    'length'
+
     """
 
     def __init__(self, unit, physical_types: Union[str, Set[str]]):
@@ -45,13 +107,26 @@ class PhysicalType:
         """Return a `set` of all physical types that represent a `Unit`."""
         return self._as_set
 
+    @property
+    def as_string(self) -> str:
+        """
+        Return a string representation of the physical type(s).
+        """
+        if len(self.as_set) == 1:
+            return list(self.as_set)[0]
+        else:
+            return "{" + str(sorted(self.as_set))[1:-1] + "}"
+
     def _get_physical_type_id(self):
+        # To match unit API
         return self._physical_type_id
 
     def __eq__(self, other):
         """Return `True` if `other` represents a physical type"""
         if isinstance(other, str):
-            return other in self.as_set
+            return other in self.as_set or other.replace("_", " ") in self.as_set
+        elif isinstance(other, set):
+            return other.issubset(self.as_set)
         elif isinstance(other, PhysicalType):
             return self._get_physical_type_id() == other._get_physical_type_id()
         else:
@@ -65,18 +140,23 @@ class PhysicalType:
 
     def __repr__(self):
         if len(self.as_set) == 1:
-            return [p for p in self.as_set][0]
+            return repr(self.as_string)
         else:
-            return str(self.as_set)
+            return self.as_string
 
     def __str__(self):
-        return self.__repr__()
+        return self.as_string
 
     @staticmethod
     def _identify_unit_from_unit_or_physical_type(obj):
+        """
+        If a unit is passed in, return that unit.  If a physical type
+        is passed in, return a unit that corresponds to that physical
+        type.
+        """
         if isinstance(obj, core.UnitBase):
             return obj
-        elif isinstance(obj, _PhysicalType):
+        elif isinstance(obj, PhysicalType):
             return obj._unit
         else:
             raise TypeError("Expecting a unit or a physical type")
@@ -86,9 +166,17 @@ class PhysicalType:
         new_unit = self._unit * other_unit
         return new_unit.physical_type
 
+    def __rmul__(self, other):
+        return self.__mul__(other)
+
     def __truediv__(self, other):
         other_unit = self._identify_unit_from_unit_or_physical_type(other)
         new_unit = self._unit / other_unit
+        return new_unit.physical_type
+
+    def __rtruediv__(self, other):
+        other_unit = self._identify_unit_from_unit_or_physical_type(other)
+        new_unit = other_unit / self._unit
         return new_unit.physical_type
 
     def __pow__(self, power):
@@ -97,47 +185,68 @@ class PhysicalType:
         return (self._unit ** power).physical_type
 
 
-def def_physical_type(unit, name):
+def def_physical_type(unit: core.UnitBase, name: Union[str, Set[str]]):
     """
-    Adds a new physical unit mapping.
+    Add a mapping between a unit and the corresponding physical type(s).
 
     Parameters
     ----------
-    unit : `~astropy.units.UnitBase` instance
-        The unit to map from.
+    unit : u.Unit
+        The unit to be represented by the physical type.
 
-    name : str
-        The physical type of the unit.
+    name : `str` or `set` of `str`
+        A `str` representing the name of the physical type of the unit,
+        or a `set` containing strings that represent one or more names
+        of physical types.
     """
-    r = unit._get_physical_type_id()
-    if r in _physical_unit_mapping:
+
+    if name in ("unknown", {"unknown"}):
+        raise ValueError("Unable to uniquely define an unknown physical type")
+
+    physical_type_id = unit._get_physical_type_id()
+    if physical_type_id in _physical_unit_mapping:
         raise ValueError(
-            f"{r!r} ({name!r}) already defined as {_physical_unit_mapping[r]!r}"
+            f"{physical_type_id!r} ({name!r}) already defined as "
+            f"{_physical_unit_mapping[physical_type_id]!r}"
         )
-    _physical_unit_mapping[r] = name
-    _unit_physical_mapping[name] = r
+
+    if isinstance(name, str):
+        name = {name}
+
+    _physical_unit_mapping[physical_type_id] = PhysicalType(unit, name)
+
+    for physical_type in name:
+        _unit_physical_mapping[physical_type] = physical_type_id
 
 
-def get_physical_type(unit):
+def get_physical_type(unit: core.UnitBase) -> PhysicalType:
     """
-    Given a unit, returns the name of the physical type it represents.
-    If it represents an unknown physical type, `"unknown"` is returned.
+    Return a representation of the physical type of a unit.
 
     Parameters
     ----------
     unit : `~astropy.units.UnitBase` instance
-        The unit to get the physical type for
+        The unit to look up
 
     Returns
     -------
-    physical_type : str
-        The name of the physical type, or `"unknown"` if not known.
+    physical : PhysicalType
+        A representation of the physical type(s) of the unit.
     """
-    r = unit._get_physical_type_id()
-    return _physical_unit_mapping.get(r, "unknown")
+    if not isinstance(unit, core.UnitBase):
+        raise TypeError("The input to get_physical_type must be a unit.")
+
+    physical_type_id = unit._get_physical_type_id()
+
+    known_unit = physical_type_id in _physical_unit_mapping
+
+    if known_unit:
+        return _physical_unit_mapping[physical_type_id]
+    else:
+        return PhysicalType(unit, "unknown")
 
 
-for unit, name in [
+units_and_physical_types = [
     (core.Unit(1), "dimensionless"),
     (si.m, "length"),
     (si.m ** 2, "area"),
@@ -155,7 +264,7 @@ for unit, name in [
     (imperial.deg_F, "temperature"),
     (si.N, "force"),
     (si.J, "energy"),
-    (si.Pa, "pressure"),
+    (si.Pa, {"pressure", "energy density"}),
     (si.W, "power"),
     (si.kg / si.m ** 3, "mass density"),
     (si.m ** 3 / si.kg, "specific volume"),
@@ -200,5 +309,7 @@ for unit, name in [
     (cgs.statampere, "electrical current (ESU)"),
     (cgs.Biot, "electrical current (EMU)"),
     (cgs.abcoulomb, "electrical charge (EMU)"),
-]:
-    def_physical_type(unit, name)
+]
+
+for unit, physical_type in units_and_physical_types:
+    def_physical_type(unit, physical_type)